pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]

[project]
name = "rdmo-catalog-rag"
description = ""
version = '1.0.0'
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "Jochen Klar", email = "main@jochenklar.de"},
]
requires-python = ">=3.11"
classifiers = [
  "Environment :: Web Environment",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "langchain",
  "langchain-community",
  "langchain-openai",
  "langchain-chroma",
  "unstructured",
  "python-dotenv",
  "chromadb",
]

[project.optional-dependencies]
dev = [
    "build",
    "pre-commit",
    "ruff",
    "twine",
]

[project.urls]
repository = "https://github.com/jochenklar/rdmo-catalog-rag.git"

[project.scripts]
rdmo-catalog-rag = "rdmo_catalog_rag.cli:main"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "DJ",   # flake8-django
  "E",    # pycodestyle
  "F",    # pyflakes
  "I",    # isort
  "PGH",  # pygrep-hooks
  "PT",   # flake8-pytest-style
  "RUF",  # ruff
  "UP",   # pyupgrade
  "W",    # pycodestyle
  "YTT",  # flake8-2020
  "G",    # flake8-logging-format
  "INT",  # flake8-gettext
  "Q"     # flake8-quotes
]
ignore = [
  "B006",   # mutable-argument-default
  "B007",   # unused-loop-control-variable
  "B018",   # useless-expression
  "G004",   # logging-f-string
  "RUF012", # mutable-class-default
]

[tool.ruff.lint.isort]
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "local-folder"
]
